Rename fileUploader to match file naming conventions.
